fix(avatar): 修复max属性不生效的问题
