+++ src/utils/render-tnode.ts
@@ -1,10 +1,11 @@
-import { h, ComponentPublicInstance, VNode, isVNode } from 'vue';
+import { h, ComponentPublicInstance, VNode, isVNode, Fragment } from 'vue';
 import isEmpty from 'lodash/isEmpty';
 import isString from 'lodash/isString';
 import isFunction from 'lodash/isFunction';
 import isObject from 'lodash/isObject';
 import camelCase from 'lodash/camelCase';
 import kebabCase from 'lodash/kebabCase';
+import isArray from 'lodash/isArray';
 
 export interface JSXRenderContext {
   defaultNode?: VNode | string;
@@ -25,6 +26,24 @@ export function getDefaultNode(options?: OptionsType) {
   }
 
   return defaultNode;
+}
+
+export function getChildren(content: VNode[]) {
+  const childList: VNode[] = [];
+  const innerGetChildren = (content: VNode[]) => {
+    if (!isArray(content)) return;
+    content.forEach((item: VNode) => {
+      if (item.children && isArray(item.children)) {
+        if (item.type !== Fragment) return;
+        innerGetChildren(item.children as VNode[]);
+      } else {
+        childList.push(item);
+      }
+    });
+    return childList;
+  };
+
+  return innerGetChildren(content);
 }
 
 export function getParams(options?: OptionsType) {
